% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgbcurves.r
\name{rgbcurves}
\alias{rgbcurves}
\title{Sensor sensitivities of the eye and of different cameras}
\usage{
rgbcurves(device = "eye")
}
\arguments{
\item{device}{from which to retrieve sensor's sensitivities. It can be any of the
devices listed in the description, that is, \code{agilent},
\code{foveonx3}, \code{kodak}, \code{nikond1}, \code{sony}, 
or \code{eye} (default).}
}
\description{
Obtain camera sensors' sensitivities. You can choose sensors for
\itemize{
  \item\code{agilent} Agilent CMOS sensor array from a Concord EyeQ digital camera
  \item\code{foveonx3} Foveon X3 sensor array from a Sigma SD9 digital camera
  \item\code{kodak} Kodak CCD sensor array from a Kodak DCS-460 digital camera
  \item\code{nikond1} a CCD sensor array from a Nikon D1 digital camera
  \item\code{sony} Sony CCD sensor array from a Hewlett-Packard digital camera
  \item\code{eye} the cone photoreceptors of the human eye
}
}
